make site manifest human readable.
There is no need to minify the json. It is impossible to review changes
when it is all scrunched on one line. This change was generated by running
`jq --tab site.webmanifest | sponge site.webmanifest` and you can verify
that by running:

$ diff -s <(git show  HEAD:kdlp.underground.software/site.webmanifest) \
          <(git show HEAD^:kdlp.underground.software/site.webmanifest | jq --tab)
Files /dev/fd/63 and /dev/fd/62 are identical